+++ serde_arrow/src/internal/array_builder.rs
@@ -1,7 +1,9 @@
 use serde::Serialize;
 
+use marrow::array::Array;
+
 use crate::internal::{
-    arrow::Array, error::Result, schema::SerdeArrowSchema, serialization::OuterSequenceBuilder,
+    error::Result, schema::SerdeArrowSchema, serialization::OuterSequenceBuilder,
 };
 
 /// Construct arrays by pushing individual records
@@ -55,12 +57,17 @@ pub struct ArrayBuilder {
 }
 
 impl ArrayBuilder {
-    /// Construct an array build from an [`SerdeArrowSchema`]
+    /// Construct an array builder from an [`SerdeArrowSchema`]
     pub fn new(schema: SerdeArrowSchema) -> Result<Self> {
         Ok(Self {
             builder: OuterSequenceBuilder::new(&schema)?,
             schema,
         })
+    }
+
+    /// Construct an array builder from `marrow` fields
+    pub fn from_marrow(_fields: Vec<marrow::datatypes::Field>) -> Result<Self> {
+        todo!()
     }
 }